Fix pre-existing beartype annotation errors surfaced at runtime
- correlation.py: Fix coords shape "1 n_edges ps ps 2" → "1 n_edges 2 ps ps"
  (coords are channel-first after permute in dpvo.py)
- correlation.py: Fix CorrLayer.forward return type from 3D to variadic
  (CUDA kernel returns 6D correlation volume)
- ba.py: Same coords shape fix
- blocks.py: Fix SoftAgg.forward ix param from "batch edges" (2D) to
  "edges" (1D) — ix is a 1D index tensor
- dpvo.py: Remove incorrect corr1/corr2 3D annotations (actual shape is 6D)
- dpvo.py: Fix DPVO.__call__ image param Float → UInt8
- scatter_utils.py: Fix _expand_index return type and local var from
  Float to Int (index tensors are int64)

+++ packages/mini-dpvo/mini_dpvo/scatter_utils.py
@@ -16,7 +16,7 @@ from jaxtyping import Float, Int
 from torch import Tensor
 
 
-def _expand_index(index: Int[Tensor, "n"], src: Float[Tensor, "*shape"], dim: int) -> Float[Tensor, "*shape"]:
+def _expand_index(index: Int[Tensor, "n"], src: Float[Tensor, "*shape"], dim: int) -> Int[Tensor, "*shape"]:
     """Broadcast a 1-D index tensor so it can be used with ``scatter_add_``.
 
     ``scatter_add_`` requires the index tensor to have the same number of
@@ -34,7 +34,7 @@ def _expand_index(index: Int[Tensor, "n"], src: Float[Tensor, "*shape"], dim: in
     """
     view: list[int] = [1] * src.dim()
     view[dim] = index.shape[0]
-    expanded: Float[Tensor, "*shape"] = index.view(*view).expand_as(src)
+    expanded: Int[Tensor, "*shape"] = index.view(*view).expand_as(src)
     return expanded
